+++ test/fixtures/amp-bind-integrations.html
@@ -1,0 +1,45 @@
+<!doctype html>
+<html ⚡>
+<head>
+  <meta charset="utf-8">
+  <title>AMP #0</title>
+  <link rel="canonical" href="amps.html" >
+  <meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1">
+  <style amp-custom>
+    h1 {color: red}
+    body {
+      max-width: 527px;
+      font-family: Arial;
+    }
+    amp-extended-sample.i-amphtml-element {
+      display: block;
+    }
+  </style>
+  <style amp-boilerplate>body{-webkit-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-moz-animation:-amp-start 8s steps(1,end) 0s 1 normal both;-ms-animation:-amp-start 8s steps(1,end) 0s 1 normal both;animation:-amp-start 8s steps(1,end) 0s 1 normal both}@-webkit-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-moz-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-ms-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@-o-keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}@keyframes -amp-start{from{visibility:hidden}to{visibility:visible}}</style><noscript><style amp-boilerplate>body{-webkit-animation:none;-moz-animation:none;-ms-animation:none;animation:none}</style></noscript>
+  <script src="/dist/amp.js"></script>
+  <script custom-element="amp-carousel" src="/dist/v0/amp-carousel-0.1.max.js"></script>
+</head>
+
+<body>
+  <p>P TAG TEST</p>
+  <button on="tap:AMP.setState(boundText='hello world')" id="boundTextButton">hello world!</button>
+  <p id="textElement" [text]="boundText">unbound</p>
+
+  <p>CAROUSEL TEST</p>
+  <button on="tap:AMP.setState(selectedSlide=1)" id="goToSlide1Button">slide 1!</button>
+  <p id="slideNum" [text]="selectedSlide">0</p>
+  <amp-carousel width="300" height="100" type="slides"  id="carousel" on="slideChange:AMP.setState(selectedSlide=event.index)" [slide]="selectedSlide">
+    <amp-img 
+    src=https://lh3.googleusercontent.com/5rcQ32ml8E5ONp9f9-Rf78IofLb9QjS5_0mqsY1zEFc=w300-h200-no
+    height="200" width="200"></amp-img>
+    <amp-img 
+    src=https://lh3.googleusercontent.com/5rcQ32ml8E5ONp9f9-Rf78IofLb9QjS5_0mqsY1zEFc=w300-h200-no
+    height="200" width="200"></amp-img>
+    <amp-img 
+    src=https://lh3.googleusercontent.com/5rcQ32ml8E5ONp9f9-Rf78IofLb9QjS5_0mqsY1zEFc=w300-h200-no
+    height="200" width="200"></amp-img>
+  </amp-carousel>
+
+
+</body>
+</html>
